[GGB-107] 필드 type 변경

+++ src/main/java/com/ggb/graduationgoodbye/domain/member/controller/MemberJoinRequest.java
@@ -6,7 +6,7 @@ public record MemberJoinRequest(
     String nickname,
     String address,
     String gender,
-    Long age,
+    Integer age,
     String phone
 ) {
